feat: discard the old configuration - 3
