+++ server/controller/grpc/synchronizer/service.go
@@ -17,7 +17,6 @@
 package synchronize
 
 import (
-	"strings"
 	"time"
 
 	api "github.com/deepflowio/deepflow/message/trident"
@@ -65,12 +64,7 @@ func (s *service) Sync(ctx context.Context, in *api.SyncRequest) (*api.SyncRespo
 }
 
 func (s *service) Push(r *api.SyncRequest, in api.Synchronizer_PushServer) error {
-	processName := r.GetProcessName()
-	if strings.HasPrefix(processName, "trident") || strings.HasPrefix(processName, "deepflow-agent") {
-		s.vTapEvent.Push(r, in)
-	} else {
-		s.tsdbEvent.Push(r, in)
-	}
+	s.tsdbEvent.Push(r, in)
 	return nil
 }
